AddToBasketModal quantity field limitation added.

+++ purchases/models.py
@@ -192,7 +192,7 @@ class InvoiceLine(models.Model):
     """ Model contains InvoiceLines for Purchases model """
     product = models.ForeignKey(Product, verbose_name=_('Goods'), on_delete=models.PROTECT)
     purchase = models.ForeignKey(Purchase, verbose_name=_('Purchase'), on_delete=models.PROTECT)
-    quantity = models.PositiveIntegerField(_('Amount'), default=1)
+    quantity = models.PositiveSmallIntegerField(_('Amount'), default=1)
     units = models.CharField(_('Units'), max_length=16, default=_('pcs.'))
     unit_price = models.DecimalField(_('Unit price'), max_digits=8, decimal_places=2, default=0)
     warehouse = models.ForeignKey(Warehouse, verbose_name=_('Warehouse'), blank=True, null=True,
